refactor: async constructor is an anti-pattern, use init instead
